[JIRA SONIC-18230] - Fixed IP Unnumbered config when 2 Loopback interfaces exists
The key to query Loopback interfaces were fetching
all the configured Loopback interfaces and counting
IPv4 address of all these. Hence when there was two
Loopback interface even with single IP configured
was failing IP Unnumbered configurations.
The fix is to query only the input Loopback address
and count all the IPv4 addresses configured on it.

Change-Id: Iefe8d8bc5719d9cafb36f175be504cc91d15d3f5