Allow non-root users to run CLISH
Prior to this change, the pipestr class would try to create the file in
the current working directory, which works file for the root user, but
fails for a non-root user.

This change works around the issue by creating a separate file that is
unique to each user in the /tmp directory, which has write access for
all users.

This has been tested with a few commands when running CLISH as admin
user to ensure that the commands work. The commit itself won't change
anything except the file where the pipe string is saved.