Make CLI actioners communicate with REST over Unix socket
Prior to this change, the REST server would spawn a TCP listener on port
8443, and require the CLI to use certificates to authenticate to the
REST server. This adds overhead (TLS and TCP), which can be avoided by
using Unix domain sockets. The advantage of using UDS is that the REST
server can determine the user at the other end of the socket connection
with standard library calls.

This means that the CLI commands can be much more performant, as they no
longer have to use HTTPS over TCP, but can use unencrypted HTTP over
UDS.

This change adds the new CLI socket authentication handler in the
authentication middleware, which determines the user at the far end of
the socket connection. The listener for the CLI is also updated to
listen on both TCP port 8443 and UDS.

It is still necessary for the server to listen on 8443, since the STP
actioner uses the Swagger generated client, which does not support Unix
domain sockets. Until the Swagger clients are removed, we must maintain
the listener on 8443.