[ham] Fix passwd command when modifying user
Translib always calls HAM's `usermod` function to either create or
modify an existing user. The `usermod` function creates the user if it
doesn't already exist, and calls separate functions to modify the saved
password and roles if it exists. However, the `passwd` function has a
bug in that it passes the hashed password directly to the `system`
command without quoting it. As a result, the shell interpreter treats it
as a set of variables to be expanded. Therefore, a hashed password like
`$6$ibQi$5rpcmxycQwIy4LD73T6pYdDBpq1cowqgVHH9fZx3YgvhlDD5LWZYpB4ooy/nZxV9mQHhzp/zo4xmHQNktk2ML0`
would be parsed by the shell as `/nZxV9mQHhzp/zo4xmHQNktk2ML0`, since
the remaining fragments `$...` would be treated as undefined (and empty)
shell variables. The usermod utility would not perform any sanity
checking on the parameter and would save the password as is to the
shadow file, thereby leaving the user to assume that the last password
command didn't succeed.

This change ensures that the hashed password is quoted within single
quotes before being executed by the shell. This has been tested with the
command `username admin password Sonic@Dell role admin` to ensure that
the password is saved correctly to the shadow file.