[JIRA SONIC-19913] VxLAN OC Yang Fix
Reverting the change of disabling the VxLAN sub-tree deletion xfmr code. This breaks the VxLAN OC
support and hence needs to be restored.

The top level interface container delete is blocked at other place as well and hence the original
issue (for which this is commented) is not able to be verified. (Apparently crash).

Test Case(s)
Executed all the OC Yang operations (Especially the Deletion of vxlan interface, vxlan src-vtep,
vlan/vni mapping deletion) at all levels and ensured it works as before.

Change-Id: I5e22e95e78f0ba15fd5876989420b82ef7302b28