Fetch user roles from translib
This change forces the REST server to attempt to fetch the list of user
roles from translib, which maintains it in a cache of the User table. If
the lookup fails for any reason, it falls back to using the group
membership to determine the user role(s).

Once HAM takes care of writing the roles to the User table, then we can
remove the fallback logic and rely exclusively on ConfigDB.