[JIRA SONIC-26747] - Fix show spanning-tree klish command
During GET, interface details are populated to a MAP
when converting DB to internal structures. Again when
converting back to OC structure, the DB is read again
instead of interface Map. This may lead to inconsistency
of interface details in both operations while STP is just enabled.
Fix is to read the details from the map while converting to OC structures.
Also in the front end, instead of accessing interface name
from 'state', the interface name is accessed via 'name'
which is the key which has to be present.

Change-Id: Id7d26e27a9cbdda5416ec27b3b59350512487e4e