[JIRA SONIC-28761] These changes enhance the logging.
Have not been able to reproduce this issue and
there is no techsupport with the defect.  I also cannot see any
failure path in the code for "no snmp-server host" that can
produce this type of "%Error: Transaction Failure".  It almost
seems as is there was an underlying MF connection reset or failure
of some kind.

These changes enhance the logging.  The intent is to ask for it
to be reproduced again, and to get the techsupport with it.

If you have any ideas on what caused the original failure then
please let me know.

Change-Id: I31f31aba3798731f0f1b5654d18858256500cce9