[JIRA SONIC-31130] Validation checks for VTEP SIP unconfigure and reconfigure
Checks are added for no source-ip to verify the presence of mappings.
Check is also added to ensure SIP can be set only when there is no SIP configured.
Checks are also made to ensure that the EVPN_NVO object is not deleted when the mappings are not present so as to not leave it in an inconsistent state.

Change-Id: I667f90bd666b30f08593e8343394cb52daab535d