[JIRA SONIC-40468] Fix return code evaluation in the rest_cl plugin
The clish_restcl plugin is returning non-zero return code when a
POST operation is performed even though the http return code is
success 200. The rest_cl() API is determining the return code
to be non-zero if the response body is non-zero. However for a
POST operation, the response body may or may not be zero.

The proposed fix is to return non-zero error code only if the
response body is non-zero and the http response code is not 200.

Change-Id: If662278899a9e73f2ba3c0a69c1eb4dcb9afc0f4