[JIRA SONIC-39456] [JIRA SONIC-39453] [JIRA SONIC-38821] : Fix show arp/neigh actioner script and DbToYang Neigh transformer for subIntfs
Show and Rest was working before with name=Ethernet64.12345 as generic code was converting it to Eth64.12345 with yeilded expected results.
This format/flow was not right and once the generic code was changed, we started to observe errors.

UT:
Verified ARP/Neighbors over Phy/Po SubInterface with dyanmic/static entries in show + intf/ip/mac filters.

Change-Id: Icec880e190f83e6ebaaf30ec7c8a7dc55103b7ac